// src/generated/models/CertificateDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CertificateDto = {
    participantId?: string;
    firstname?: string;
    lastname?: string;
    quizathonTitle?: string;
    quizathonDate?: string;
    position?: number;
    qrCode?: string;
};